Define project structure
The idea is to proper structure Steel using modern CMake
to divide the project in:
- steel: the library (which for now is an application);
- steel_tests: the module that will be used for testing Steel;
- sample: a demo using Steel as a library

The objective is to build and install Steel as a library into
an output directory and then use it for tests (so we don't need
to rebuild it). That is not done yet, but this is a good starting
point.